Refactor factory placement
