+++ packages/examples/cvat/exchange-oracle/src/handlers/job_creation/__init__.py
@@ -1,3 +1,3 @@
-from src.handlers.job_creation.factory import create_task
+from src.handlers.job_creation.handlers import create_task
 
 __all__ = ["create_task"]
